springboot-mvc: mvc with thymeleaf
Before going into the details of each file. Let me first explain the root or the heart of the project i.e. Thymeleaf (pronounced as 'time-leaf'). It is a template engine to work with HTML, XML along with java. It creates the ease for data passing in-between html pages as well as in-between html to controller classes and so on.  I'll explain further details about Thymeleaf during my per file explanation.

DemoController.java: it is a separate controller mainly to explain the 'Model' class. Model is the kind of container which is used to move data in-between html pages and controller class. it uses two params, one is 'key' and other one is 'value'. here it is returning the string 'welcomePage'. And here comes the Thymeleaf. Thymeleaf look for welcomePage.html in template folder and will display it in the browser.

welcomePage.html: this is the html page opens from the DemoController.java class. Here you can see the '${TheDate}' parameter in 'p' tag. this TheDate is the same key that I defined in DemoController method while adding the value in Model.

HelloWorldController.Java: Here this file is linked with 'helloworld-form.html' and explains how we can use the values providing in html form in java classes. the postMapping 'processForm' get a call from the html form. defined in helloworld-form.html. those V2 and V3 are just different ways to process the values. in 'V2', I am using HttpServletRequest, and extacting the request parameters from it. In 'V3', I am using an annotation 'RequestParam' which helps us directly providing the value without using the HttpServletRequest.
helloworld-form.html: here you can see that 'studentName' key that we were using in HelloWorldController to extract the value.

StudentController: This is the class which is using all the concepts that I explain in previous classes. And Additionally, if you want to populate some values from application.properties file, you can use @Value annotation to linked that field with the data defined in application.properties. Other than that, this class is using the full MVC. Where Student.Java is the Model, StudentController is the Controller class and sutdent-form and student-confirmation are Views.

student-form.html: Here I am creating a form and saving the information in Student Object using DataBinding. Here 'th:object= ${student}' we are linking the student object that we added in Model object in our controller class. And method 'post' indicates that it would send a post request upon form submission. Within the form, as you can see, '*{firstName}' I am accessing the object fields directly. We can also use '${student.firstName}'. but the former approach represents a more organised pattern.   Here I am also showing other different type  of html fields and how can we linked them via dataBinding.